Added services for each table, added the item controller, and reworked the model mapping. The server is able to be compiled

+++ Spring Boot/project-1/src/main/java/com/skillstorm/demo/repositories/itemRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.skillstorm.demo.models.item;
 
 @Repository
-public interface itemRepository extends JpaRepository<item, Integer>{
+public interface itemRepository extends JpaRepository<item, Long>{
 
 }
